+++ Common.Mod.Generator/Specs/ConfigEntrySpec.cs
@@ -5,6 +5,7 @@ public record ConfigEntrySpec
     public string Name { get; set; } = null!;
     public ConfigEntryTypeSpec Type { get; set; }
     public object? DefaultValue { get; set; } = null;
+    public string? Enum { get; set; } = null;
     public string? Nested { get; set; } = null;
     public string? Label { get; set; } = null;
     public string? Description { get; set; } = null;