+++ Common.Mod.Generator/Specs/ConfigEntryTypeSpec.cs
@@ -10,5 +10,6 @@ public enum ConfigEntryTypeSpec
     Float,
     Double,
     String,
+    Enum,
     Nested
 }